增加部分空文件夹和文件的逻辑判断，增加cors解决前后端跨域问题

+++ app/main.py
@@ -24,9 +24,12 @@ def file_builder(file_list):
 
 
 def clean_files():
-    os.remove("combined_PDF.pdf")
-    shutil.rmtree("tmp")
-    os.mkdir("tmp")
+    if os.path.exists("combined_PDF.pdf"):
+        os.remove("combined_PDF.pdf")
+    if os.path.exists("tmp"):
+        shutil.rmtree("tmp")
+        os.mkdir("tmp")
+    return True
 
 # 这段是借鉴的，目前是把一张a4纸上面放两个pdf，具体程度待测试
 def pdf_merge(pdfs, output):
